feat: Renderizando o carrossel e aplicando o efeito de foco
